requirements_model.txt
torchvision==0.16.2
pillow==10.2.0
torch >= 2.1.0
joblib==1.3.2
uvicorn==0.25.0
fastapi==0.108.0
hydra-core==1.3.0
wandb==0.16.1
python-multipart==0.0.6
dvc==3.38.1
"dvc[gdrive]"
scikit-learn==1.3.2